adding in register via development because of github merge infinite cycle issue
when merging register to dev, there was a bug where the merge could not be completed, yet would continually ask me after the error to finalize the merge, resulting in a infinite loop that gives an error each time. i deleted register and backed up my files locally, swapped to dev and carefully made sure that the files i brought in didnt recreate the error.